Major rework
Had to rework destroy mode, the threads were not properly dying. I ended
up changing to a BackgroundWorker to get the functionality we wanted.
Also had to change the search mode listbox to a combobox to get the
functionality we wanted. It is now populated properly as well. Used
background worker to thread various activities with the turret as well.